python/requirements.txt
sseclient==0.0.27
paho-mqtt==1.6.1